MDEV-16315: NOT FOUND condition not triggered in stored aggregate functions
Aggregate function has:
 - initialization code
 - code that is run for every row
 - code that generates a result
For an empty table we don't run the middle part but for custom aggregate we are running
the middle part also.
Fixed this by only allowing the generation of result. As soon as we encounter the first FETCH GROUP NEXT ROW
instruction with empty tables, we should exit and tell the handler that there is nothing to fetch and return
the result.